Address comments v1
