+++ test-frame-common/src/main/java/io/skodjob/testframe/wait/Wait.java
@@ -4,12 +4,24 @@
  */
 package io.skodjob.testframe.wait;
 
+import io.skodjob.testframe.LoggerUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.time.Duration;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import java.util.function.BooleanSupplier;
 
 public class Wait {
+    private static final Logger LOGGER = LoggerFactory.getLogger(Wait.class);
 
     /**
      * For every poll (happening once each {@code pollIntervalMs}) checks if supplier {@code ready} is true.
@@ -98,5 +110,65 @@ public class Wait {
                 return;
             }
         }
+    }
+
+    private static final ExecutorService EXECUTOR = Executors.newCachedThreadPool(new ThreadFactory() {
+        final ThreadFactory defaultThreadFactory = Executors.defaultThreadFactory();
+
+        @Override
+        public Thread newThread(Runnable r) {
+            Thread result = defaultThreadFactory.newThread(r);
+            result.setDaemon(true);
+            return result;
+        }
+    });
+
+    /**
+     * For every poll (happening once each {@code pollIntervalMs}) checks if supplier {@code ready} is true.
+     * If yes, the wait is closed. Otherwise, waits another {@code pollIntervalMs} and tries again.
+     * Once the wait timeout (specified by {@code timeoutMs} is reached and supplier wasn't true until that time,
+     * runs the {@code onTimeout} (f.e. print of logs, showing the actual value that was checked inside {@code ready}),
+     * and finally throws {@link WaitException}.
+     *
+     * @param description information about on what we are waiting
+     * @param pollIntervalMs poll interval in milliseconds
+     * @param timeoutMs timeout specified in milliseconds
+     * @param ready {@link BooleanSupplier} containing code, which should be executed each poll, verifying readiness
+     *                                     of the particular thing
+     */
+    public static CompletableFuture<Void> untilAsync(String description, long pollIntervalMs, long timeoutMs, BooleanSupplier ready) {
+        LOGGER.info("Waiting for {}", description);
+        long deadline = System.currentTimeMillis() + timeoutMs;
+        CompletableFuture<Void> future = new CompletableFuture<>();
+        Executor delayed = CompletableFuture.delayedExecutor(pollIntervalMs, TimeUnit.MILLISECONDS, EXECUTOR);
+        Runnable r = new Runnable() {
+            @Override
+            public void run() {
+                boolean result;
+                try {
+                    result = ready.getAsBoolean();
+                } catch (Exception e) {
+                    future.completeExceptionally(e);
+                    return;
+                }
+                long timeLeft = deadline - System.currentTimeMillis();
+                if (!future.isDone()) {
+                    if (!result) {
+                        if (timeLeft >= 0) {
+                            if (LOGGER.isTraceEnabled()) {
+                                LOGGER.trace("{} not ready, will try again ({}ms till timeout)", description, timeLeft);
+                            }
+                            delayed.execute(this);
+                        } else {
+                            future.completeExceptionally(new TimeoutException(String.format("Waiting for %s timeout %s exceeded", description, timeoutMs)));
+                        }
+                    } else {
+                        future.complete(null);
+                    }
+                }
+            }
+        };
+        r.run();
+        return future;
     }
 }
