+++ test-frame-kubernetes/src/main/java/io/skodjob/testframe/resources/JobResource.java
@@ -30,6 +30,15 @@ public class JobResource implements NamespacedResourceType<Job> {
     }
 
     /**
+     * Get specific client for resoruce
+     * @return specific client
+     */
+    @Override
+    public MixedOperation<?, ?, ?> getClient() {
+        return client;
+    }
+
+    /**
      * Creates specific {@link Job} resource in Namespace specified by user
      *
      * @param namespaceName Namespace, where the resource should be created
