+++ test-frame-kubernetes/src/main/java/io/skodjob/testframe/resources/RoleBindingResource.java
@@ -30,6 +30,15 @@ public class RoleBindingResource implements NamespacedResourceType<RoleBinding> 
     }
 
     /**
+     * Get specific client for resoruce
+     * @return specific client
+     */
+    @Override
+    public MixedOperation<?, ?, ?> getClient() {
+        return client;
+    }
+
+    /**
      * Creates specific {@link RoleBinding} resource
      *
      * @param resource {@link RoleBinding} resource
